all: Use let-chains when possible
Now that we moved to 1.88.0 MSRV, we can use that syntax.

Run `cargo +1.89.0 clippy --all-targets --all-features --tests --no-deps -pcoreutils -- -D warnings`
until it became happy.

This was useful, as `--fix` doesn't directly work on the command above:
ls src/uu/ | xargs -I{} cargo +1.89.0 clippy --all-targets --all-features --tests --no-deps -puu_{} --fix --allow-dirty -- -D warnings

+++ src/uu/mkdir/src/mkdir.rs
@@ -276,12 +276,12 @@ fn create_dir(path: &Path, is_parent: bool, config: &Config) -> UResult<()> {
 
             // Apply SELinux context if requested
             #[cfg(feature = "selinux")]
-            if config.set_selinux_context && uucore::selinux::is_selinux_enabled() {
-                if let Err(e) = uucore::selinux::set_selinux_security_context(path, config.context)
-                {
-                    let _ = std::fs::remove_dir(path);
-                    return Err(USimpleError::new(1, e.to_string()));
-                }
+            if config.set_selinux_context
+                && uucore::selinux::is_selinux_enabled()
+                && let Err(e) = uucore::selinux::set_selinux_security_context(path, config.context)
+            {
+                let _ = std::fs::remove_dir(path);
+                return Err(USimpleError::new(1, e.to_string()));
             }
 
             Ok(())
